test_drf_authorization/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
python-dotenv==1.0.0
django-phonenumber-field==7.3.0
drf-spectacular==0.27.2